move ion_sound_speed tests to test_speeds.py
