+++ plasmapy/formulary/tests/test_parameters.py
@@ -46,14 +46,7 @@ from plasmapy.formulary.parameters import (
     wuh_,
 )
 from plasmapy.particles import Particle
-from plasmapy.particles.exceptions import InvalidParticleError
-from plasmapy.utils.exceptions import (
-    PhysicsError,
-    PhysicsWarning,
-    PlasmaPyFutureWarning,
-    RelativityError,
-    RelativityWarning,
-)
+from plasmapy.utils.exceptions import PlasmaPyFutureWarning
 from plasmapy.utils.pytest_helpers import assert_can_handle_nparray
 
 B = 1.0 * u.T
@@ -145,221 +138,6 @@ class Test_mass_density:
 
 # Assertions below that are in CGS units with 2-3 significant digits
 # are generally from the NRL Plasma Formulary.
-
-
-class Test_Ion_Sound_Speed:
-    r"""Test the ion_sound_speed function in parameters.py."""
-
-    @pytest.mark.parametrize(
-        "args, kwargs, expected, isclose_kw",
-        [
-            (
-                (),
-                {
-                    "T_i": 1.3232 * u.MK,
-                    "T_e": 1.831 * u.MK,
-                    "ion": "p",
-                    "gamma_e": 1,
-                    "gamma_i": 3,
-                },
-                218816.06086407552 * (u.m / u.s),
-                {},
-            ),
-            (
-                (1.831 * u.MK, 1.3232 * u.MK, "p"),
-                {},
-                218816.06086407552 * (u.m / u.s),
-                {},
-            ),  # Test that function call without keyword argument works correctly
-            (
-                (),
-                {
-                    "T_i": 1.3232 * u.MK,
-                    "T_e": 1.831 * u.MK,
-                    "n_e": n_e,
-                    "k": k_1,
-                    "ion": "p",
-                    "gamma_e": 1,
-                    "gamma_i": 3,
-                },
-                218816.06086407552 * (u.m / u.s),
-                {},
-            ),
-            (
-                (),
-                {
-                    "T_i": 1.3232 * u.MK,
-                    "T_e": 1.831 * u.MK,
-                    "n_e": n_e,
-                    "k": k_2,
-                    "ion": "p",
-                    "gamma_e": 1,
-                    "gamma_i": 3,
-                },
-                552.3212936293337 * (u.m / u.s),
-                {},
-            ),
-            (
-                (),
-                {
-                    "T_i": 0.88 * u.MK,
-                    "T_e": 1.28 * u.MK,
-                    "n_e": n_e,
-                    "k": 0 * u.m ** -1,
-                    "ion": "p",
-                    "gamma_e": 1.2,
-                    "gamma_i": 3.4,
-                },
-                193328.52857788358 * (u.m / u.s),
-                {},
-            ),
-            (
-                (),
-                {"T_i": T_i, "T_e": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p+"},
-                ion_sound_speed(T_i=T_i, T_e=0 * u.K, n_e=n_e, k=k_1, ion="p+").value
-                * (u.m / u.s),
-                {},
-            ),
-            (
-                (),
-                {
-                    "T_e": 1.2e6 * u.K,
-                    "T_i": 0 * u.K,
-                    "n_e": n_e,
-                    "k": 0 * u.m ** -1,
-                    "z_mean": 0.8,
-                    "ion": "p",
-                },
-                89018.09 * (u.m / u.s),
-                {"atol": 0.0, "rtol": 1e-6},
-            ),  # testing for user input z_mean
-        ],
-    )
-    def test_values(self, args, kwargs, expected, isclose_kw):
-        assert np.isclose(ion_sound_speed(*args, **kwargs), expected, **isclose_kw)
-
-    # case when Z=1 is assumed
-    # assert ion_sound_speed(T_i=T_i, T_e=T_e, ion='p+') == ion_sound_speed(T_i=T_i, T_e=T_e,
-    # ion='H-1')
-
-    @pytest.mark.parametrize(
-        "kwargs1, kwargs2, _warning",
-        [
-            ({"T_i": T_i, "T_e": T_e, "n_e": n_e, "ion": "p"}, {}, PhysicsWarning),
-            ({"T_i": T_i, "T_e": T_e, "k": k_1, "ion": "p"}, {}, PhysicsWarning),
-            ({"T_i": 5e11 * u.K, "T_e": 0 * u.K, "ion": "p"}, {}, RelativityWarning),
-            (
-                {"T_e": 1.2e6, "T_i": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p"},
-                {"T_e": 1.2e6 * u.K, "T_i": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p"},
-                u.UnitsWarning,
-            ),
-            (
-                {"T_i": 1.3e6, "T_e": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p"},
-                {"T_i": 1.3e6 * u.K, "T_e": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p"},
-                u.UnitsWarning,
-            ),
-        ],
-    )
-    def test_warns(self, kwargs1, kwargs2, _warning):
-        with pytest.warns(_warning):
-            val = ion_sound_speed(**kwargs1)
-            if kwargs2 != {}:
-                val == ion_sound_speed(**kwargs2)
-
-    @pytest.mark.parametrize(
-        "args, kwargs, _error",
-        [
-            (
-                (),
-                {
-                    "T_i": T_i,
-                    "T_e": T_e,
-                    "n_e": n_e,
-                    "k": k_1,
-                    "ion": "p",
-                    "gamma_i": np.inf,
-                },
-                RelativityError,
-            ),
-            (
-                (),
-                {
-                    "T_i": np.array([5, 6, 5]) * u.K,
-                    "T_e": np.array([3, 4]) * u.K,
-                    "n_e": np.array([5, 6, 5]) * u.m ** -3,
-                    "k": np.array([3, 4]) * u.m ** -3,
-                    "ion": "p",
-                },
-                u.UnitTypeError,
-            ),
-            ((5 * u.T), {"ion": "p"}, TypeError),  # Is this test right??????
-            ((), {"ion": "p"}, TypeError),
-            (
-                (),
-                {"T_i": T_i, "T_e": 0 * u.K, "gamma_i": 0.9999, "ion": "p"},
-                PhysicsError,
-            ),
-            (
-                (),
-                {"T_i": T_i, "T_e": 0 * u.K, "gamma_e": 0.9999, "ion": "p"},
-                PhysicsError,
-            ),
-            (
-                (),
-                {"T_i": T_i, "T_e": 0 * u.K, "gamma_e": "sdjklsf", "ion": "p"},
-                TypeError,
-            ),
-            (
-                (),
-                {"T_i": T_i, "T_e": 0 * u.K, "gamma_i": "fsdfas", "ion": "p"},
-                TypeError,
-            ),
-            ((), {"T_i": T_i, "T_e": 0 * u.K, "ion": "cupcakes"}, InvalidParticleError),
-            ((), {"T_i": -np.abs(T_i), "T_e": 0 * u.K, "ion": "p"}, ValueError),
-            (
-                (),
-                {"T_i": T_i, "T_e": 0 * u.K, "n_e": -np.abs(n_e), "k": k_1, "ion": "p"},
-                ValueError,
-            ),
-            (
-                (),
-                {"T_i": T_i, "T_e": 0 * u.K, "n_e": n_e, "k": -np.abs(k_1), "ion": "p"},
-                ValueError,
-            ),
-            ((), {"T_i": 5e19 * u.K, "T_e": 0 * u.K, "ion": "p"}, RelativityError),
-            (
-                (),
-                {"T_i": 5 * u.A, "T_e": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p"},
-                u.UnitTypeError,
-            ),
-            (
-                (),
-                {"T_i": T_negarr, "T_e": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p"},
-                ValueError,
-            ),
-            (
-                (),
-                {"T_e": T_negarr, "T_i": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p"},
-                ValueError,
-            ),
-        ],
-    )
-    def test_raises(self, args, kwargs, _error):
-        with pytest.raises(_error):
-            ion_sound_speed(*args, **kwargs)
-
-    @pytest.mark.parametrize(
-        "kwargs",
-        [
-            ({"T_i": T_nanarr, "T_e": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p"}),
-            ({"T_e": T_nanarr, "T_i": 0 * u.K, "n_e": n_e, "k": k_1, "ion": "p"}),
-        ],
-    )
-    def test_nan_values(self, kwargs):
-        np.isnan(ion_sound_speed(**kwargs)[1])
-
-    def test_handle_nparrays(self):
-        assert_can_handle_nparray(ion_sound_speed)
 
 
 def test_thermal_pressure():
@@ -462,8 +240,6 @@ def test_Bohm_diffusion():
     "alias, parent",
     [
         (rho_, mass_density),
-        (va_, Alfven_speed),
-        (cs_, ion_sound_speed),
         (pth_, thermal_pressure),
         (pmag_, magnetic_pressure),
         (ub_, magnetic_energy_density),
@@ -603,11 +379,13 @@ def test_parameters_aliases(alias, parent):
             Alfven_speed,
             speeds.Alfven_speed,
         ),
+        ({"B": 0.4 * u.T, "density": 1e18 * u.m ** -3, "ion": "He+"}, va_, speeds.va_),
         (
-            {"B": 0.4 * u.T, "density": 1e18 * u.m ** -3, "ion": "He+"},
-            va_,
-            speeds.va_,
+            {"T_e": 58000 * u.K, "T_i": 12000 * u.K, "ion": "He+"},
+            ion_sound_speed,
+            speeds.ion_sound_speed,
         ),
+        ({"T_e": 58000 * u.K, "T_i": 12000 * u.K, "ion": "He+"}, cs_, speeds.cs_),
     ],
 )
 def test_deprecated(kwargs, deprecated_func, parent):
